refactor: Move fs internal structures to new internal fs directory
