+++ std/node/_fs/_fs_dirent.ts
@@ -1,4 +1,4 @@
-import { notImplemented } from "./_utils.ts";
+import { notImplemented } from "../_utils.ts";
 
 export default class Dirent {
   constructor(private entry: Deno.FileInfo) {}
